+++ test/object_group_test.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'dart:ui';
 
 import 'package:test/test.dart';
 import 'package:tiled/tiled.dart';
@@ -27,7 +28,11 @@ void main() {
     });
 
     test('sets color', () {
-      expect(objectGroup.color, equals('#555500'));
+      expect(objectGroup.colorHex, equals('#555500'));
+      expect(
+        objectGroup.color,
+        equals(Color(int.parse('0xff555500'))),
+      );
     });
 
     test('sets opacity', () {